refactor: 트리뷰, 메모카드 그리드 폴더 수정
