--- apps/client/src/features/tree-view/index.ts
@@ -1,6 +1,0 @@
-export { default as TreeBaseMemoNode } from './ui/tree-base-memo-node/tree-base-memo-node';
-export { default as TreeCustomEdgeLabel } from './ui/tree-custom-edge-label/tree-custom-edge-label';
-export { default as TreeCustomEdgeNoLabel } from './ui/tree-custom-edge-no-label/tree-custom-edge-no-label';
-export { default as TreeCustomHandle } from './ui/tree-custom-handle/tree-custom-handle';
-export { default as TreeMemoList } from './ui/tree-memo-list/tree-memo-list';
-export { default as TreeMemoListNode } from './ui/tree-memo-list-node/tree-memo-list-node';
