+++ apps/client/src/pages/ai-results/ai-results-page.tsx
@@ -6,8 +6,7 @@ import {
   MemoListView,
   type MemoListViewHelpers,
 } from '@shared/features/memo-list-view';
-
-import { type MockMemo } from '@widgets/memo-list/types/memo';
+import { type MockMemo } from '@shared/features/memo-list-view/components/memo-list/types/memo';
 
 import { useGetAIMemo } from './apis/queries';
