+++ apps/client/src/pages/all-memo/hooks/use-all-memo.ts
@@ -1,8 +1,7 @@
 import { useCallback, useEffect, useMemo, useState } from 'react';
 
 import { SelectedMemo } from '@shared/features/ai-prompt/components/index';
-
-import { type MockMemo } from '@widgets/memo-list/types/memo';
+import { type MockMemo } from '@shared/features/memo-list-view/components/memo-list/types/memo';
 
 // 검색 관련
 interface UseMemoSearchReturn {
