+++ apps/client/src/pages/new-memo/components/memo-input/api/query-key.ts
