+++ apps/client/src/pages/new-memo/components/memo-input/index.ts
@@ -1,0 +1,1 @@
+export { default as MemoInput } from './memo-input';
