+++ apps/client/src/pages/new-memo/components/memo-input/memo-input.css.ts
