--- apps/client/src/pages/new-memo/components/memo/index.ts
@@ -1,1 +1,0 @@
-export { default as MemoInput } from './ui/memo-input/memo-input';
