+++ apps/client/src/shared/features/memo-list-view/components/memo-list/index.ts
@@ -1,0 +1,1 @@
+export { default as MemoCardGrid } from './memo-card-grid';
