+++ apps/client/src/shared/features/memo-list-view/components/memo-list/memo-card-grid.css.ts
