+++ apps/client/src/shared/features/memo-list-view/components/memo-list/types/memo.ts
