+++ apps/client/src/shared/features/memo-list-view/components/tree-view/index.ts
@@ -1,0 +1,1 @@
+export { default as TreeView } from './tree-view';
