--- apps/client/src/widgets/memo-list/index.ts
@@ -1,1 +1,0 @@
-export { default as MemoCardGrid } from './ui/memo-card-grid';
