--- apps/client/src/widgets/tree-view/index.ts
@@ -1,1 +1,0 @@
-export { default as TreeView } from './ui/tree-view';
